Put both AuthToken impls in same file
